+++ test/http-test.js
@@ -416,6 +416,16 @@ describe("httptest", () => {
 
       expect(resp.body).to.equal(JSON.stringify({ a: "b" }));
     });
+
+    it("content-type application/json returns body as json", async () => {
+      const resp = await request(app)
+        .post("/api")
+        .send({ a: "b" })
+        .expect("content-type", "application/json; charset=utf-8")
+        .expect(201);
+
+      expect(resp.body).to.deep.equal({ a: "b" });
+    });
   });
 
   describe("redirects", () => {
@@ -496,13 +506,14 @@ describe("httptest", () => {
   });
 
   describe("error", () => {
-    it("api throws is request is aborted", async () => {
+    it("api throws if request is aborted", async () => {
       try {
-        await request(app, { bodyTimeout: 50 }).get("/api/parse/204");
+        await request(app, { bodyTimeout: 50 }).get("/econnreset");
       } catch (e) {
         var err = e;
       }
 
+      expect(err).to.be.ok;
       expect(err).to.have.property("code", "ECONNRESET");
     });
   });